Place mongodb driver functions into its own file.
