+++ budgie.go
@@ -11,6 +11,9 @@ const MongoDb = "budgie"
 const MongoCollection = "expenses"
 const MongoUri = "mongodb://127.0.0.1:27017" // running this on localhost
 
+// other constants
+const default_feedback = "Press Ctrl+C to go back to home screen."
+
 func main() {
 
 	p := tea.NewProgram(createHomeScreenModel())
